fix bundle and caching
